# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Flappy Bird Backend API
  version: 1.0.0
  summary: Optional accounts, leaderboards, achievements and score sync for the Flappy Bird iOS game.
  description: |
    # Overview

    This service is the **optional** companion backend for the
    [Flappy Bird iOS game](https://github.com/hoangsonww/Flappy-Bird-Game).

    The game is fully playable without it. When a server is reachable, the app
    unlocks accounts, cloud-synced progress, global/daily/weekly leaderboards,
    friend boards, achievements and the deterministic daily challenge.

    ## Discovery handshake

    On launch the game probes a short list of candidate base URLs and calls
    `GET /v1/meta/config`. A response containing `"protocol": "flappy-bird/1"`
    switches the app into online mode; anything else keeps it offline. Nothing
    else about the app's behaviour depends on the server being present.

    ```mermaid
    sequenceDiagram
      participant Game as iOS Game
      participant API as Backend
      Game->>API: GET /v1/meta/config
      API-->>Game: 200 { protocol, capabilities, limits }
      Game->>API: POST /v1/auth/guest  { deviceId }
      API-->>Game: 201 { user, tokens }
      Game->>API: POST /v1/scores      (Bearer access token)
      API-->>Game: 201 { score, rank, unlockedAchievements }
    ```

    ## Authentication

    Two token types:

    | Token         | Lifetime            | Storage on device | Purpose                        |
...

...

...
    history.

...

...

...
    {
...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/leaderboard:
    get:
      tags: [Leaderboard]
      summary: Global leaderboard
      description: |
        One row per player, representing their best eligible run inside the
        window. Ties share a rank and are ordered by the earlier submission.
      security: []
      operationId: getLeaderboard
      parameters:
        - $ref: '#/components/parameters/ModeQuery'
        - $ref: '#/components/parameters/WindowQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: A page of the board
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeaderboardPage' }
  /v1/leaderboard/me:
    get:
      tags: [Leaderboard]
      summary: Own rank with neighbours
      description: Returns the caller's row plus the players immediately above and below.
      operationId: getOwnRank
      parameters:
        - $ref: '#/components/parameters/ModeQuery'
        - $ref: '#/components/parameters/WindowQuery'
        - name: radius
          in: query
          description: How many rows to include on each side of the caller.
          schema: { type: integer, minimum: 0, maximum: 25, default: 3 }
      responses:
        '200':
          description: Rank context
          content:
            application/json:
              schema:
                type: object
                properties:
                  window: { $ref: '#/components/schemas/LeaderboardWindow' }
                  mode: { type: string, example: all }
                  rank: { type: [integer, 'null'] }
                  total: { type: integer }
                  entry:
                    { oneOf: [{ $ref: '#/components/schemas/LeaderboardEntry' }, { type: 'null' }] }
                  neighbours:
                    type: array
                    items: { $ref: '#/components/schemas/LeaderboardEntry' }
  /v1/leaderboard/friends:
    get:
      tags: [Leaderboard]
      summary: Friends-only board
      description: Ranks the caller against the accounts they follow.
      operationId: getFriendsLeaderboard
      parameters:
        - $ref: '#/components/parameters/ModeQuery'
        - $ref: '#/components/parameters/WindowQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Friends board
          content:
            application/json:
              schema:
                type: object
                properties:
                  window: { $ref: '#/components/schemas/LeaderboardWindow' }
                  mode: { type: string }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/LeaderboardEntry' }
                  total: { type: integer }
  /v1/leaderboard/stream:
    get:
      tags: [Leaderboard]
      summary: Live board updates (SSE)
      description: |
        `text/event-stream`. Emits `connected` on subscribe, `heartbeat` every
        25 s, and `score.submitted` whenever an eligible run lands.

        ```
        event: score.submitted
        data: {"type":"score.submitted","at":"2026-03-19T09:14:02.881Z","payload":{"username":"skyhopper","score":42,"mode":"classic","rank":3}}
        ```

        Disabled when `ENABLE_SSE=false`.
      security: []
      operationId: streamLeaderboard
      responses:
        '200':
          description: An open event stream
          content:
            text/event-stream:
              schema: { type: string }
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/challenges/{date}/leaderboard:
    get:
      tags: [Challenges]
      summary: Challenge leaderboard
      security: []
      operationId: getChallengeLeaderboard
      parameters:
        - $ref: '#/components/parameters/DatePath'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Ranked entries for that date
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string, format: date }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ChallengeEntry' }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  hasMore: { type: boolean }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token from `/v1/auth/login`, `/register`, `/guest` or `/refresh`.
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Static token matching the server's `ADMIN_TOKEN`. An alternative to an admin account.
  parameters:
...
    DatePath:
      name: date
      in: path
      required: true
      description: UTC date in `YYYY-MM-DD` form.
      schema: { type: string, format: date, example: '2026-03-19' }
    ModeQuery:
      name: mode
      in: query
      description: Restrict to a single game mode. Omit for all modes combined.
      schema: { $ref: '#/components/schemas/GameMode' }
    WindowQuery:
      name: window
      in: query
      description: Time window, evaluated in UTC. Weeks start on Monday.
      schema: { $ref: '#/components/schemas/LeaderboardWindow' }
    LimitQuery:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    OffsetQuery:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  schemas:
    Username:
      type: string
      pattern: '^[a-zA-Z0-9_]{3,20}$'
      example: skyhopper
      description: 3–20 characters; letters, digits and underscore. Compared case-insensitively.
    BirdSkin:
      type: string
      enum: [classic, midnight, ember, mint, royal, glitch, aurora, phoenix]
      description: Cosmetic bird skin. Unlocked in game with coins.
    GameMode:
      type: string
      enum: [classic, endless, timeAttack, hardcore, zen, daily]
      description: |
        * `classic` — the original rules.
        * `endless` — difficulty ramps continuously.
        * `timeAttack` — score as much as possible in 60 seconds.
        * `hardcore` — narrow gaps, one life, no power-ups.
        * `zen` — no death, for practice (excluded from most boards by convention).
        * `daily` — today's deterministic challenge.
    LeaderboardWindow:
      type: string
      enum: [all, daily, weekly, monthly]
      default: all
...
    LeaderboardEntry:
      type: object
      properties:
        rank: { type: integer, minimum: 1 }
        scoreId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        username: { $ref: '#/components/schemas/Username' }
        displayName: { type: string }
        avatarSkin: { $ref: '#/components/schemas/BirdSkin' }
        country: { type: [string, 'null'] }
        score: { type: integer }
        mode: { $ref: '#/components/schemas/GameMode' }
        coins: { type: integer }
        pipesPassed: { type: integer }
        durationMs: { type: integer }
        achievedAt: { type: string, format: date-time }
    LeaderboardPage:
      type: object
      properties:
        window: { $ref: '#/components/schemas/LeaderboardWindow' }
        mode: { type: string, example: all }
        items:
          type: array
          items: { $ref: '#/components/schemas/LeaderboardEntry' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        hasMore: { type: boolean }
        generatedAt: { type: string, format: date-time }
...
    ChallengeEntry:
      type: object
      properties:
        rank: { type: integer }
        date: { type: string, format: date }
        userId: { type: string, format: uuid }
        username: { $ref: '#/components/schemas/Username' }
        displayName: { type: string }
        avatarSkin: { $ref: '#/components/schemas/BirdSkin' }
        score: { type: integer }
        submittedAt: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - validation_failed
                - unauthorized
                - invalid_credentials
                - token_expired
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - payload_too_large
                - unprocessable
                - service_unavailable
                - internal_error
            message: { type: string }
            details: { description: Structured context; shape depends on `code`. }
            requestId: { type: string, description: Matches the `X-Request-Id` response header. }
  responses:
...
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...